Add checks to ensure scriptis are executed in user dropbox

+++ gigadb/app/tools/files-metadata-console/scripts/filesMetaToDb.sh
@@ -13,7 +13,6 @@ APP_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 WORKING_DIR=$(pwd)
 
 DOI=$1
-
 if [ -z "${DOI}" ];then
   if [[ $(uname -n) =~ compute ]]; then
     echo -e "Usage: /usr/local/bin/filesMetaToDb <DOI>\n"
@@ -23,8 +22,36 @@ if [ -z "${DOI}" ];then
   exit 1;
 fi
 
-if [[ $(uname -n) =~ compute ]];then
+#######################################
+# Check doi.md5 and doi.filesizes files
+# exist in current working directory
+# Globals:
+#   WORKING_DIR
+# Arguments:
+#   None
+#######################################
+function check_files_exist() {
+  # Check current directory contains doi.md and doi.filesizes
+  if ! test -f "${WORKING_DIR}/${DOI}.md"; then
+    err "A ${DOI}.md file is required in this directory"
+    exit 1
+  fi
+  if ! test -f "${WORKING_DIR}/${DOI}.filesizes"; then
+    err "A ${DOI}.filesizes file is required in this directory"
+    exit 1
+  fi
+}
+
+if [[ $(uname -n) =~ compute ]]; then
   . /home/centos/.bash_profile
+  
+  # Check current directory is a user dropbox
+  if [[ ! "${WORKING_DIR}" == *"/share/dropbox/user"* ]]; then
+    err "filesMetaToDb script should only be used in a user directory located at /share/dropbox"
+    exit 1
+  fi
+
+  check_files_exist
 
   echo -e "Updating md5 checksum values as file attributes for ${DOI}"
   docker run --rm -v "${WORKING_DIR}":/gigadb/app/tools/files-metadata-console/metadata "registry.gitlab.com/${GITLAB_PROJECT}/production-files-metadata-console:${GIGADB_ENV}" ./yii update/md5-values --doi="${DOI}"
@@ -32,6 +59,14 @@ if [[ $(uname -n) =~ compute ]];then
   docker run --rm -v "${WORKING_DIR}":/gigadb/app/tools/files-metadata-console/metadata "registry.gitlab.com/${GITLAB_PROJECT}/production-files-metadata-console:${GIGADB_ENV}" ./yii update/file-sizes --doi="${DOI}"
 
 else
+  # Check developer is at gigadb-website/gigadb/app/tools/files-metadata-console/tests/_data/dropbox/user
+  if [[ ! "${WORKING_DIR}" == *"gigadb-website/gigadb/app/tools/files-metadata-console/tests/_data/dropbox/user"* ]]; then
+    err "filesMetaToDb.sh script should only be used in a files-metadata-console/tests/_data/dropbox/user* directory in dev environment"
+    exit 1
+  fi
+
+  check_files_exist
+
   echo -e "Updating md5 checksum values as file attributes for ${DOI}"
   docker-compose run --rm -v "${WORKING_DIR}":/gigadb/app/tools/files-metadata-console/metadata files-metadata-console ./yii update/md5-values --doi="${DOI}"
   echo -e "Updating file sizes for $DOI"
